+++ src/debug.c
@@ -259,6 +259,7 @@ int desensamblar_instruccion(const Chunk *c, int offset, FILE *out) {
         case OP_ASEGURAR_CADENA: return instruccion_simple("OP_ASEGURAR_CADENA", offset, out);
         case OP_REPR:            return instruccion_simple("OP_REPR", offset, out);
         case OP_ASIGNAR_REBANADA: return instruccion_simple("OP_ASIGNAR_REBANADA", offset, out);
+        case OP_FORMATO_F_SPEC: return instruccion_const_corta(c, offset, out);
         case OP_LONGITUD:        return instruccion_simple("OP_LONGITUD", offset, out);
         case OP_ES_TUPLA:        return instruccion_simple("OP_ES_TUPLA", offset, out);
         case OP_ES_LISTA:        return instruccion_simple("OP_ES_LISTA", offset, out);